Remove _constructor property and initialize w/init
The recommended way to subclass pandas data structures is with:

```py
@property
def _constructor()
    return SubClassedProperty
```
But that seems to create a ton of issues with the BlockManager. I
suspect that the better solution there is to re-implement getters and
setters for the subclassed frame, but working with simple inheritance
seems to be getting the job done for now.