Update GraphFrame node/rel attrs to nframe/rframe
Changes node and rels attributes to a more semantic nframe and rframe
attribute call, and updates the head method to return a GraphFrame,
allowing a user to string additional calls onto `head`, e.g.:
```gf.head(5).nframe```
returns a NodeFrame of the top 5 nodes in the GraphFrame. Whereas
```gf.head(5).nodes```
will now return the associated Node objects with the top 5 indicies in
the GraphFrame.